server/requirements.txt
flask==2.3.3
transformers==4.41.1
torch==2.3.0
torchaudio==2.3.0
soundfile==0.12.1
numpy==1.26.4
Flask
assemblyai
python-dotenv